Actors now die when they have 0 health. Can make ranged attacks from anywhere, these will "simulate" projectiles, with damage being reduced when traversing obstacles.

+++ GridFight/GridFight/GameObject.h
@@ -24,7 +24,7 @@ public:
 	}
 	FloatRect GetGlobalBounds() { return icon.getGlobalBounds(); }
 	OBJECT_TYPE GetType() { return type; }
-	void SetScale(int scale) { icon.setScale(scale, scale); }
+	void SetScale(float scale) { icon.setScale(scale, scale); }
 
 protected:
 	Sprite icon;
